+++ tests/regression/test_environments.py
@@ -7,7 +7,8 @@ import acoular as ac
 import numpy as np
 import pytest
 from pytest_cases import parametrize_with_cases
-from test_environments_cases import Environments, Flows
+
+from tests.cases.test_environments_cases import Environments, Flows
 
 GRID = ac.RectGrid3D(x_min=-0.2, x_max=0.2, y_min=-0.2, y_max=0.2, z_min=0.5, z_max=0.9, increment=0.2)
 MICS = ac.MicGeom(mpos_tot=((0.5, 0.5, 0), (0, 0, 0), (-0.5, -0.5, 0)))
@@ -16,7 +17,7 @@ MICS = ac.MicGeom(mpos_tot=((0.5, 0.5, 0), (0, 0, 0), (-0.5, -0.5, 0)))
 @pytest.mark.parametrize('grid', [GRID.gpos], ids=['RectGrid3D'])
 @parametrize_with_cases('flow', cases=Flows)
 def test_flow(snapshot, grid, flow):
-    """Performs snapshot testing with snapshot fixture from pytest-regtest
+    """Performs snapshot testing with snapshot fixture from pytest-regtest.
 
     To overwrite the collected snapshots, run:
 
@@ -41,7 +42,7 @@ def test_flow(snapshot, grid, flow):
 @pytest.mark.parametrize('grid', [GRID.gpos], ids=['RectGrid3D'])
 @parametrize_with_cases('env', cases=Environments)
 def test_environment(snapshot, grid, mics, env):
-    """Performs snapshot testing with snapshot fixture from pytest-regtest
+    """Performs snapshot testing with snapshot fixture from pytest-regtest.
 
     To overwrite the collected snapshots, run:
