+++ tests/regression/test_generator.py
@@ -4,12 +4,12 @@
 """Snapshot tests for all generators."""
 
 from pytest_cases import parametrize_with_cases
-from test_generator_cases import Generators
+from tests.cases.test_generator_cases import Generators
 
 
 @parametrize_with_cases('obj', cases=Generators)
 def test_generators(snapshot, obj):
-    """Performs snapshot testing with snapshot fixture from pytest-regtest
+    """Performs snapshot testing with snapshot fixture from pytest-regtest.
 
     Uses the generator cases defined in class Generators from test_generator_cases.py
 
@@ -26,7 +26,6 @@ def test_generators(snapshot, obj):
     obj : instance of acoular.base.Generator
         Generator instance to be tested (cases from Generators)
     """
-
     gen = obj.result(num=1)
     result = next(gen)  # Trigger calculation
     snapshot.check(result, rtol=5e-5, atol=5e-8)  # uses numpy.testing.assert_allclose
