+++ tests/unittests/test_sdinput.py
@@ -7,7 +7,7 @@ import acoular as ac
 import pytest
 
 
-@pytest.fixture
+@pytest.fixture()
 def sounddevice_samples_generator(sounddevice_properties):
     """Fixture to create an instance of ac.SoundDeviceSamplesGenerator."""
     _, max_input_channels, device_index, num = sounddevice_properties
@@ -17,8 +17,8 @@ def sounddevice_samples_generator(sounddevice_properties):
 
 def test_default_sounddevice_properties(sounddevice_properties):
     """Test that the default samplerate and input channels from sounddevice are
-    correctly transferred to ac.SoundDeviceSamplesGenerator class."""
-
+    correctly transferred to ac.SoundDeviceSamplesGenerator class.
+    """
     default_samplerate, max_input_channels, device_index, num = sounddevice_properties
     sdev = ac.SoundDeviceSamplesGenerator(device=device_index)
     assert sdev.sample_freq == default_samplerate
@@ -68,12 +68,12 @@ def test_set_sample_freq(sounddevice_properties):  # noqa ARG001 (uses fixture f
     assert sdev.sample_freq == new_sample_freq
 
 
-def test_import_error(mock_have_no_lib):  # noqa ARG001 (uses fixture from conftest.py)
+def test_import_error(_mock_have_no_lib):  # noqa ARG001 (uses fixture from conftest.py)
     """Test that ac.SoundDeviceSamplesGenerator raises ImportError if sounddevice is not available.
 
     Parameters
     ----------
-    mock_have_no_lib : pytest.fixture
+    _mock_have_no_lib : pytest.fixture
         Mock behaviour: sounddevice library does not exist.
     """
     del ac.config.__dict__['_traits_cache_have_sounddevice']  # remove cached property
